# src/swagger/swagger.json
{
    "openapi": "3.0.0",
    "components": {
...
        "schemas": {
...
            "EpisodeCacheDefinition": {
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "url": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "cacheRefreshed": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "required": ["id", "url", "name"],
                "type": "object",
                "additionalProperties": true
            },
            "Record_string.string_": {
                "properties": {},
                "additionalProperties": {
                    "type": "string"
                },
                "type": "object",
                "description": "Construct a type with a set of properties K of type T"
            },
            "QualityProfile": {
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "quality": {
                        "type": "string"
                    },
                    "sizeFactor": {
                        "type": "number",
                        "format": "double"
                    }
                },
                "required": ["id", "name", "quality", "sizeFactor"],
                "type": "object",
                "additionalProperties": true
            },
            "AppType": {
                "enum": ["SONARR", "RADARR", "PROWLARR", "SABNZBD", "NZBGET"],
                "type": "string"
            },
            "App": {
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "type": {
                        "$ref": "#/components/schemas/AppType"
                    },
                    "name": {
                        "type": "string"
                    },
                    "url": {
                        "type": "string"
                    },
                    "api_key": {
                        "type": "string"
                    },
                    "username": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "priority": {
                        "type": "number",
                        "format": "double"
                    },
                    "iplayarr": {
                        "properties": {
                            "useSSL": {
                                "type": "boolean"
                            },
                            "port": {
                                "type": "number",
                                "format": "double"
                            },
                            "host": {
                                "type": "string"
                            }
                        },
                        "required": ["useSSL", "port", "host"],
                        "type": "object"
                    },
                    "download_client": {
                        "properties": {
                            "port": {
                                "type": "number",
                                "format": "double"
                            },
                            "api_key": {
                                "type": "string"
                            },
                            "host": {
                                "type": "string"
                            },
                            "name": {
                                "type": "string"
                            },
                            "id": {
                                "type": "number",
                                "format": "double"
                            }
                        },
                        "required": ["id"],
                        "type": "object"
                    },
                    "indexer": {
                        "properties": {
                            "priority": {
                                "type": "number",
                                "format": "double"
                            },
                            "api_key": {
                                "type": "string"
                            },
                            "url": {
                                "type": "string"
                            },
                            "name": {
                                "type": "string"
                            },
                            "id": {
                                "type": "number",
                                "format": "double"
                            }
                        },
                        "required": ["priority", "id"],
                        "type": "object"
                    }
                },
                "required": ["id", "type", "name", "url", "iplayarr"],
                "type": "object",
                "additionalProperties": true
            },
            "AppFeature": {
                "enum": [
                    "api_key",
                    "callback",
                    "download_client",
                    "indexer",
                    "prowlarr_download_client",
                    "prowlarr_indexer",
                    "username_password",
                    "priority"
                ],
                "type": "string"
            },
            "Record_AppType.AppFeature-Array_": {
                "properties": {
                    "SONARR": {
                        "items": {
                            "$ref": "#/components/schemas/AppFeature"
                        },
                        "type": "array"
                    },
                    "RADARR": {
                        "items": {
                            "$ref": "#/components/schemas/AppFeature"
                        },
                        "type": "array"
                    },
                    "PROWLARR": {
                        "items": {
                            "$ref": "#/components/schemas/AppFeature"
                        },
                        "type": "array"
                    },
                    "SABNZBD": {
                        "items": {
                            "$ref": "#/components/schemas/AppFeature"
                        },
                        "type": "array"
                    },
                    "NZBGET": {
                        "items": {
                            "$ref": "#/components/schemas/AppFeature"
                        },
                        "type": "array"
                    }
                },
                "required": ["SONARR", "RADARR", "PROWLARR", "SABNZBD", "NZBGET"],
                "type": "object",
                "description": "Construct a type with a set of properties K of type T"
            }
        },
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "name": "X-API-Key",
                "in": "header"
            }
        }
    },
    "info": {
        "title": "iplayarr",
        "version": "1.0.0",
        "license": {
            "name": "ISC"
        },
        "contact": {}
    },
    "paths": {
...
        "/api/offSchedule": {
            "get": {
                "operationId": "GetAll",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/EpisodeCacheDefinition"
                                    },
                                    "type": "array"
                                }
                            }
                        }
                    }
                },
                "tags": ["EpisodeCacheDefinitions"],
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "parameters": []
            },
            "post": {
                "operationId": "Create",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EpisodeCacheDefinition"
                                }
                            }
                        }
                    }
                },
                "tags": ["EpisodeCacheDefinitions"],
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EpisodeCacheDefinition"
                            }
                        }
                    }
                }
            },
            "put": {
                "operationId": "Update",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EpisodeCacheDefinition"
                                }
                            }
                        }
                    }
                },
                "tags": ["EpisodeCacheDefinitions"],
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EpisodeCacheDefinition"
                            }
                        }
                    }
                }
            },
            "delete": {
                "operationId": "Delete",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "boolean"
                                }
                            }
                        }
                    }
                },
                "tags": ["EpisodeCacheDefinitions"],
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/offSchedule/refresh": {
            "post": {
                "operationId": "Refresh",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "boolean"
                                }
                            }
                        }
                    }
                },
                "tags": ["EpisodeCacheDefinitions"],
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EpisodeCacheDefinition"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/config/hiddenSettings": {
            "get": {
                "operationId": "HiddenSettings",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Record_string.string_"
                                }
                            }
                        }
                    }
                },
                "tags": ["Settings"],
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "parameters": []
            }
        },
        "/api/config/qualityProfiles": {
            "get": {
                "operationId": "QualityProfiles",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/QualityProfile"
                                    },
                                    "type": "array"
                                }
                            }
                        }
                    }
                },
                "tags": ["Settings"],
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "parameters": []
            }
        },
        "/api/apps": {
            "get": {
                "operationId": "GetAll",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/App"
                                    },
                                    "type": "array"
                                }
                            }
                        }
                    }
                },
                "tags": ["Apps"],
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "parameters": []
            },
            "post": {
                "operationId": "Create",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/App"
                                }
                            }
                        }
                    }
                },
                "tags": ["Apps"],
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/App"
                            }
                        }
                    }
                }
            },
            "put": {
                "operationId": "Update",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/App"
                                }
                            }
                        }
                    }
                },
                "tags": ["Apps"],
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/App"
                            }
                        }
                    }
                }
            },
            "delete": {
                "operationId": "Delete",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "boolean"
                                }
                            }
                        }
                    }
                },
                "tags": ["Apps"],
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/apps/types": {
            "get": {
                "operationId": "Types",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Record_AppType.AppFeature-Array_"
                                }
                            }
                        }
                    }
                },
                "tags": ["Apps"],
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "parameters": []
            }
        },
        "/api/apps/test": {
            "post": {
                "operationId": "Test",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "boolean"
                                }
                            }
                        }
                    }
                },
                "tags": ["Apps"],
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/App"
                            }
                        }
                    }
                }
            }
        },
        "/api/apps/updateApiKey": {
            "post": {
                "operationId": "UpdateApiKey",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "boolean"
                                }
                            }
                        }
                    }
                },
                "tags": ["Apps"],
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "parameters": []
            }
        }
    },
    "servers": [
        {
            "url": "/"
        }
    ]
}